+++ src/asp/Core/StereoSettings.cc
@@ -131,8 +131,9 @@ namespace asp {
        "Remove IP within this percentage from the outer edges of an image pair (integer percent).")
       ("normalize-ip-tiles", po::bool_switch(&global.ip_normalize_tiles)->default_value(false)->implicit_value(true),
        "Individually normalize tiles used for IP detection.")
-      ("ip-inlier-factor",          po::value(&global.ip_inlier_factor)->default_value(1.0/15.0),
-       "A higher factor will result in more interest points, but perhaps also more outliers, and a bigger search range.")
+      ("ip-inlier-factor",          po::value(&global.ip_inlier_factor)->default_value(0.2),
+       "A higher factor will result in more interest points, but perhaps also more outliers, "
+       "and a bigger search range.")
       ("ip-uniqueness-threshold",          po::value(&global.ip_uniqueness_thresh)->default_value(0.8),
        "Min percentage distance between closest and second closest IP descriptors, a larger value allows more IP matches.")
       ("ip-nodata-radius",          po::value(&global.ip_nodata_radius)->default_value(4),
@@ -315,7 +316,7 @@ namespace asp {
                      "Extend SGM calculation to this distance to increase accuracy at tile borders.")
       ("sgm-search-buffer",        po::value(&global.sgm_search_buffer)->default_value(Vector2i(4,4),"4 4"),
                      "Search range expansion for SGM down stereo pyramid levels.  Smaller values are faster, but greater change of blunders.")
-      ("corr-memory-limit-mb",     po::value(&global.corr_memory_limit_mb)->default_value(4*1024),
+      ("corr-memory-limit-mb",     po::value(&global.corr_memory_limit_mb)->default_value(5*1024),
        "Keep correlation memory usage (per tile) close to this limit.  Important for SGM/MGM.")
       ("correlator-mode", po::bool_switch(&global.correlator_mode)->default_value(false)->implicit_value(true),
        "Function as an image correlator only (including with subpixel refinement). Assume no cameras, aligned input images, and stop before triangulation, so at filtered disparity.")